requirements.txt
python-telegram-bot>=21,<23
python-dotenv>=1.0
